Renamed utils module to compiler
